Fix tests for games app

+++ games/util/pagination.py
@@ -2,8 +2,8 @@ def get_page_range(total_pages, current_page, max_range=5):
     if total_pages <= max_range + 2:
         return range(1, total_pages + 1)
     pages = []
-    start_page = max(1, current_page - max_range / 2)
-    end_page = min(total_pages, current_page + max_range / 2)
+    start_page = max(1, current_page - int(max_range / 2))
+    end_page = min(total_pages, current_page + int(max_range / 2))
     if start_page > 1:
         pages.append(1)
     if start_page > 2:
